chore: remove dead code and fix workflow batch mode
- remove dead vision module (blur, exposure, optical_flow) and unused re-exports in imu/insta360/output/selection/video
- fix colmap camera model mapping (OPENCV_FISHEYE 5 vs 6) with param validation and tests
- fix colmap database pragma handling and rig duplicate handling
- fix geometry overlap and orientation tests (rad vs deg)
- improve scene-mask: colmap-compatible masks by default, per-file timing, turbo preset, vision helper via swiftc
- fix workflow.py: batch mode by default (--interactive to pause), robust pause handling for \r/^M, stdin DEVNULL, total timing
- update .gitignore for room/dataset/__pycache__

+++ src/insta360/metadata.rs
@@ -46,10 +46,10 @@ fn probe_file(path: &Path) -> Result<StreamInfo> {
     };
     let duration = ictx.duration();
     // ffmpeg duration is in AV_TIME_BASE (microseconds), -1 if unknown
-    let duration_us = if duration == ffmpeg_next::ffi::AV_NOPTS_VALUE as i64 || duration < 0 {
+    let duration_us = if duration == ffmpeg_next::ffi::AV_NOPTS_VALUE || duration < 0 {
         // fallback to stream duration
         let dur_ts = stream.duration();
-        if dur_ts == ffmpeg_next::ffi::AV_NOPTS_VALUE as i64 || dur_ts < 0 {
+        if dur_ts == ffmpeg_next::ffi::AV_NOPTS_VALUE || dur_ts < 0 {
             0
         } else {
             let tb = stream.time_base();
